Add buzzer tones to startup
Short monotone at start, 4 second delay, then two short notes. The first
tone confirms the code was downloaded OK or that it is OK to attach to
the serial port. The delay is a window to download new code in case the
current one will crash. The second sound (double-tone) means we're off
to the races!