finger_filter: Improve filtering + small reorganization
Use array of Ema's instead of just short_ema & long_ema.
Add is_stable that makes sure the Ema's are close to each other and
going in a downward trend before recognizing a down event.
Make a timeout on the down event after which we start updating Ema's
again. Otherwise, a sudden shift downward that never ends would leave
the "button" in a down state indefinitely.
Fix a bug that caused a long press to appear as multiple 1/2-second
presses.