Fix button trigger condition - first working
Calculate a diffs array to store the difference between the sample and
each successively longer EMA. Use this to evaluate stability and whether
all EMAs are in the right order (i.e. both short & long-term direction
is downward).

The diffs are published in the misc[] field of the RangeButton ROS
message. An extra two entries are tacked on to that with extra debug
data.