RangeButton: Reset VL6180X driver if hardware gets stuck
Add timeout logic to the Adafruit_VL6180X driver so that we cannot get
stuck indefinitely waiting for the hardware to be ready.

HACK: Add a driver_status field that the client can use to determine
that the driver / hardware got stuck during a ranging operation.

Split out driver reset and internal state reset into helper functions
(moved from ::init()), and use these to reset everything if the hardware
locked up.  This includes cycling the enable pin to reset the HW.