Enable hardware watchdog.
There have been some instances of the system locking up, so let's
recover automatically when that happens. Right now, we'll only service
the dog when we get VALID sensor readings (i.e. not "max distance / no
return"). So, we set the timeout as long as possible (16 seconds) to
prevent a rapid boot loop when sensor is not in range of anything.

The reason we don't service when we're getting "max dist" readings is
that perhaps it could be a failure mode of the sensor rather than
physical reality.